Тесты Selenium/requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
pytest==7.4.3
pytest-html==4.0.2
allure-pytest==2.13.2
pytest-xdist==3.5.0
pytest-parallel==0.1.1
requests==2.31.0